+++ AVInput/CHANGELOG.md
@@ -16,6 +16,10 @@ All notable changes to this RDK Service will be documented in this file.
 
 * For more details, refer to [versioning](https://github.com/rdkcentral/rdkservices#versioning) section under Main README.
 
+## [1.2.1] - 2023-04-12
+### Fixed
+- Fixed IARM_Bus_UnRegisterEventHandler  call to IARM_Bus_RemoveEventHandler
+
 ## [1.2.0] - 2023-04-03
 ### Added
 - Typo in RPC response to AVInput call org.rdk.AVInput.getSupportedGameFeatures