+++ SystemServices/CHANGELOG.md
@@ -16,6 +16,10 @@ All notable changes to this RDK Service will be documented in this file.
 
 * For more details, refer to [versioning](https://github.com/rdkcentral/rdkservices#versioning) section under Main README.
 
+## [1.3.1] - 2023-04-12
+### Fixed
+- Fixed IARM_Bus_UnRegisterEventHandler  call to IARM_Bus_RemoveEventHandler
+
 ## [1.3.0] - 2023-04-03
 ### Added
 - Added new API getWakeupSrcConfiguration and updated setWakeupSrcConfiuration API to use across all platforms.